requirements.txt
tk
matplotlib
Pillow